Commit generated native bindings, add a proj property switch to enable regeneration (off by default)
